Refactor tokenBalances into utils file
